add `historicalDataCollection` option to the health data component, restructure the components into separate files
